Updated file paths so everything is followedCompany instead of saved company.
