+++ client/src/components/CompanyChart.js
@@ -1,29 +1,20 @@
 import React, { useState, useEffect } from "react";
-import { Typography, Container, Box } from "@material-ui/core";
+import { Box, Typography } from "@material-ui/core";
 import { Doughnut } from "react-chartjs-2";
-import { useQuery, useMutation } from '@apollo/client';
-import { GET_ME } from '../utils/queries';
 
 export default function CustomCharts({ data }) {
-
   const [chartData, setChartData] = useState();
-  const [score, setScore] = useState([])
-  // setScore.push(data.environmentScore, data.socialScore, data.governanceScore );
-// setScore
 
   useEffect(() => {
     let newScore = []
-    newScore.push(data.environmentScore, data.socialScore, data.governanceScore );
-    setScore(newScore)
+    newScore.push(data.environmentScore, data.socialScore, data.governanceScore);
 
     const newChart = {
-
       labels: [
-        'Enivormental Score',
+        'Environmental Score',
         'Social Score',
         'Governance Score'
       ],
-  
       datasets: [{
         label: 'something',
         data: newScore,
@@ -34,37 +25,31 @@ export default function CustomCharts({ data }) {
         ],
         hoverOffset: 4
       }]
-  
-    } 
-      setChartData(newChart)
-
-  }, [setScore, data])
-
- 
+    }
+    setChartData(newChart)
+  }, [data])
 
   return (
     <>
-      
       <Box sx={{
         width: 400,
         height: 200,
       }}>
-        <h1>{data.companyName}</h1>
+        <h1>{data.companyName ? data.companyName : "Loading"}</h1>
         <Typography variant="h6">ESG Scores</Typography>
-      <Doughnut
-        data={chartData}
-        options={{
-          maintainAspectRatio: true,
-          title: {
-            display: true,
-            text: "ESG Scores",
-            fontSize: 25 
-          },
-          legend: { display: true, position: "bottom" }
-        }}
+        <Doughnut
+          data={chartData}
+          options={{
+            maintainAspectRatio: true,
+            title: {
+              display: true,
+              text: "ESG Scores",
+              fontSize: 25
+            },
+            legend: { display: true, position: "bottom" }
+          }}
         />
       </Box>
-
     </>
   );
-}
+};
