+++ server/models/Company.js
@@ -2,7 +2,7 @@ const { Schema } = require('mongoose');
 
 const goalsSchema = require('./Goals');
 
-// This is a subdocument schema, it won't become its own model but we'll use it as the schema for the User's `savedCompanies` array in User.js
+// This is a subdocument schema, it won't become its own model but we'll use it as the schema for the User's `followedCompanies` array in User.js
 const companySchema = new Schema({
   // ID number given by the API.
   companyId: {
